+++ wrapper/server/mcserver/console_parser/__init__.py
@@ -14,12 +14,16 @@ class ConsoleParser:
 
     def parse(self, line):
         # Compatible with most recent versions of Minecraft server
-        r = re.search("(\[[0-9:]*\]) \[([A-z #0-9]*)\/([A-z #]*)\](.*)", line)
+        new_style = re.search("(\[[0-9:]*\]) \[([A-z #0-9]*)\/([A-z #]*)\](.*)", line)
+        old_style = re.search("([0-9-: ]*) \[([A-Z]*)\] (.*)", line)
 
         # If regex did not match, continue to prevent issues
-        if r == None:
-            return
+        if new_style:
+            self.new_style(new_style)
+        elif old_style:
+            self.old_style(old_style)
 
+    def new_style(self, r):
         log_time = r.group(1)
         server_thread = r.group(2)
         log_level = r.group(3)
@@ -150,7 +154,7 @@ class ConsoleParser:
                 "\[(.*): Teleported (.*) to (.*), (.*), (.*)\]",
                 output
             )
-            
+
             if r and not r2:
                 username = r.group(1)
                 x, y, z = r.group(2), r.group(3), r.group(4)
@@ -225,5 +229,153 @@ class ConsoleParser:
                 ": Showing new (.*) for (.*)",
                 output
             )
+            if r:
+                return False
+
+    def old_style(self, r):
+        log_time = r.group(1)
+        log_level = r.group(2)
+        output = r.group(3)
+
+        if self.mcserver.state == SERVER_STARTING:
+            # Grab server version
+            r = re.search("Starting minecraft server version (.*)", output)
+            if r:
+                server_version = r.group(1)
+                self.mcserver.server_version = server_version
+
+                for release in PROTOCOL_VERSIONS:
+                    if release["minecraftVersion"] == server_version:
+                        self.mcserver.server_version_protocol = \
+                            release["version"]
+
+            # Grab server port
+            r = re.search("Starting Minecraft server on \*:([0-9]*)", output)
+            if r:
+                server_port = r.group(1)
+
+            # Grab world name
+            r = re.search("Preparing level \"(.*)\"", output)
+            if r:
+                level_name = r.group(1)
+                self.mcserver.world = World(r.group(1))
+
+            if output == "The server will make no attempt to authenticate usernames. Beware.":
+                self.mcserver.online_mode = False
+
+            # Server started
+            if "Done" in output:
+                self.mcserver.state = SERVER_STARTED
+
+                self.mcserver.events.call("server.started")
+
+        if self.mcserver.state == SERVER_STARTED:
+            # Player Command
+            r = re.search(
+                "\[(.*): (.*)\]",
+                output
+            )
+            if r:
+                username = r.group(1)
+                command_response = r.group(2)
+
+                player = self.mcserver.get_player(username=username)
+
+                self.mcserver.events.call(
+                    "server.player.command_response",
+                    player=player,
+                    command_response=command_response
+                )
+
+            # Player Position
+            r = re.search(
+                "Teleported (.*) to (.*), (.*), (.*)",
+                output
+            )
+            r2 = re.search(
+                "\[(.*): Teleported (.*) to (.*), (.*), (.*)\]",
+                output
+            )
+
+            if r and not r2:
+                username = r.group(1)
+                x, y, z = r.group(2), r.group(3), r.group(4)
+                x, y, z = float(x), float(y), float(z)
+
+                player = self.mcserver.get_player(username=username)
+                player.position = [x, y, z]
+                player.db["position_last_updated"] = time.time()
+
+                self.mcserver.events.call(
+                    "server.player.position",
+                    player=player,
+                    x=x,
+                    y=y,
+                    z=z
+                )
+
+                return False
+
+            # Player Join
+            r = re.search("(.*) \[\/(.*):(.*)\] logged in with entity id (.*) at \((.*), (.*), (.*)\)", output)
+            if r:
+                username = r.group(1)
+                ip_address = r.group(2)
+                entity_id = r.group(4)
+                position = [
+                    float(r.group(5)),
+                    float(r.group(6)),
+                    float(r.group(7))
+                ]
+
+                try:
+                    player = self.mcserver.get_player(username=username)
+                except TypeError:
+                    try:
+                        mcuuid = self.mcserver.uuid_cache.get(username)
+                    except EOFError:
+                        mcuuid = self.mcserver.uuid_cache.get_offline_uuid(username)
+
+                    player = Player(server=self.mcserver.server, username=username, mcuuid=mcuuid)
+
+                    self.mcserver.players.append(player)
+
+                    print(mcuuid, player)
+
+                player.online = True
+                player.ip_address = ip_address
+
+                self.mcserver.events.call("server.player.join", player=player)
+
+            # Player Part
+            r = re.search("(.*) lost connection: (.*)", output)
+            if r:
+                username = r.group(1)
+                server_disconnect_reason = r.group(2)
+
+                player = self.mcserver.get_player(username=username)
+
+                if player:
+                    player.online = False
+                    self.mcserver.events.call("server.player.part", player=player)
+
+            # Chat messages
+            r = re.search("<(.*)> (.*)", output)
+            if r:
+                username, message = r.group(1), r.group(2)
+
+                player = self.mcserver.get_player(username=username)
+                self.mcserver.events.call(
+                    "server.player.message",
+                    player=player,
+                    message=message
+                )
+
+            # Misc. surpressions
+            r = re.search(
+                "Showing new (.*) for (.*)",
+                output
+            )
+
             if r:
                 return False